Added a Script we will base future work on
In this example I took Google Maps, but we will make this Script common so it will be applicable to any app. Note the comments on how we will request everything we need from the Android system. I know the comments are German but even if you cannott read them you will see whatwill happen if you follow our future commits