Only backup apk if it includes a dex file
Most of the stock manufacturer ROMs are deodexed which means that the apks in the system partition of such ROMs don't have the odexed Java Code in the apk file itself. As a result those apks can not be installed again. We won't backup those apks. We will only backup the data of this app.